Adds a cleaner way to route diy controllers and concerns refactoring
- moves shared code for QuestionForm and DiyForm into FormAttributes
concern
- moves shared code for QuestionNavigation and DiyNavigation into
ControllerNavigation concern

- (#172730228) Create protected "start your FSA" page
