+++ app/lib/question_navigation.rb
@@ -156,36 +156,6 @@ class QuestionNavigation
     Questions::SuccessfullySubmittedController,
   ].freeze
 
-  class << self
-    delegate :first, to: :controllers
+  include ControllerNavigation
 
-    def controllers
-      FLOW
-    end
-  end
-
-  delegate :controllers, to: :class
-
-  def initialize(current_controller)
-    @current_controller = current_controller
-  end
-
-  def next
-    return unless index
-
-    controllers_until_end = controllers[index + 1..-1]
-    seek(controllers_until_end)
-  end
-
-  private
-
-  def index
-    controllers.index(@current_controller.class)
-  end
-
-  def seek(list)
-    list.detect do |controller_class|
-      controller_class.show?(@current_controller.current_intake)
-    end
-  end
 end
