+++ app/views/diy/file_yourself/edit.html.erb
@@ -45,7 +45,7 @@
             Actually, I need assistance filing.
           <% end %>
 
-          <%= link_to personal_info_diy_path, class: "button button--primary text--centered button--wide spacing-above-60" do %>
+          <%= link_to next_path, class: "button button--primary text--centered button--wide spacing-above-60" do %>
             Continue
           <% end %>
         </main>
