+++ config/routes.rb
@@ -54,15 +54,13 @@ Rails.application.routes.draw do
   resources :documents, only: [:destroy]
   resources :ajax_mixpanel_events, only: [:create]
 
-  resource :diy, controller: :diy do
-    collection do
-      DiyNavigation.controllers.uniq.each do |controller_class|
-        { get: :edit, put: :update }.each do |method, action|
-          match "/#{controller_class.to_param}",
-                action: action,
-                controller: controller_class.controller_path,
-                via: method
-        end
+  scope :diy, as: :diy do
+    DiyNavigation.controllers.uniq.each do |controller_class|
+      { get: :edit, put: :update }.each do |method, action|
+        match "/#{controller_class.to_param}",
+          action: action,
+          controller: controller_class.controller_path,
+          via: method
       end
     end
   end
