+++ spec/features/diy_intake/new_diy_client_spec.rb
@@ -1,7 +1,7 @@
 require "rails_helper"
 
 RSpec.feature "Web Intake New Client wants to file on their own" do
-  scenario "new client wants to use FSA to file", :focus do
+  scenario "new client wants to use FSA to file" do
     visit "/diy/file-yourself"
 
     expect(page).to have_selector("h1", text: "File your taxes yourself!")
